+++ packages/af-utils-react-virtual-headless/jest.config.js
@@ -1,7 +1,0 @@
-export default {
-    testEnvironment: "jsdom",
-    moduleNameMapper: {
-        "src/(.*)": "<rootDir>/src/$1"
-    },
-    transform: {}
-};