+++ packages/af-utils-react-virtual-headless/src/hooks/useVirtualModel/index.ts
@@ -1,7 +1,7 @@
 import { useRef } from "react";
 import List, { ListInitialParams } from "models/List";
 
-const useVirtualModel = (params: ListInitialParams): List =>
+const useVirtualModel = (params: ListInitialParams) =>
     (useRef<List>().current ||= new List(params));
 
 export default useVirtualModel;