+++ packages/af-utils-react-virtual-headless/src/utils/assert/index.ts
@@ -1,0 +1,7 @@
+const assert = (condition: boolean, errorMessage: string) => {
+    if (!condition) {
+        throw new Error(errorMessage);
+    }
+};
+
+export default assert;