+++ packages/af-utils-react-virtual-headless/src/utils/growTypedArray/index.ts
@@ -1,0 +1,14 @@
+import FTreeArray from "models/FTreeArray";
+
+const growTypedArray = (
+    sourceArray: FTreeArray,
+    newLength: number,
+    fillValue: number
+) => {
+    const resultArray = new FTreeArray(newLength);
+    resultArray.set(sourceArray);
+    resultArray.fill(fillValue, sourceArray.length);
+    return resultArray;
+};
+
+export default growTypedArray;