+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @af-utils/react-mobx-table
 
+## 1.0.3
+
+### Patch Changes
+
+-   upgraded deps, fixed headless updating range on itemCount change, improved syntax highlighting
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.1.11
+    -   @af-utils/react-table@1.0.3
+    -   @af-utils/styled@0.0.23
+
 ## 1.0.2
 
 ### Patch Changes